Pull out scoring logic into own file
